backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-iot-device
azure-iot-hub
azure-cosmos
gunicorn